+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add trace origin to span data ([#849](https://github.com/getsentry/sentry-go/pull/849))
 - Add ability to skip frames in stacktrace ([#852](https://github.com/getsentry/sentry-go/pull/852))
 - Remove Martini integration ([#861](https://github.com/getsentry/sentry-go/pull/861))
+- Add Slog integration ([#865](https://github.com/getsentry/sentry-go/pull/865))
 
 ## 0.28.1